pyproject.toml
[project]
name = "geminicli2api"
version = "2.0.0"
description = "Universal Gemini API proxy (multi-account OAuth) + Gemini Translator CLI"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # --- Server ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "google-auth-oauthlib>=1.2.0",
    "pydantic>=2.10.0",
    "requests>=2.32.0",
    # --- Translator ---
    "httpx>=0.27.0",
    "pysrt>=1.1.2",
    "pysubs2>=1.7.0",
    "natsort>=8.4.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "anyio[trio]>=4.0.0",
    "pytest-anyio>=0.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"
